requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.10
django-angular==1.1.4
django-tables2==1.12.0
djangorestframework==3.7.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
pydot==1.2.3
pyparsing==2.2.0
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
